+++ chaincode/artifactevent-sc-javascript/lib/constants.js
@@ -1,0 +1,13 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+"use strict";
+
+const ARTIFACT_EVENT_INDEX = "artifact~event";
+const ARTIFACT_EVENT_ASSET_LABEL = "ArtifactEvent";
+
+module.exports = {
+  ARTIFACT_EVENT_ASSET_LABEL,
+  ARTIFACT_EVENT_INDEX,
+};